+++ apps/docs/src/components/actions/ActionGroup.stories.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { ActionGroup, Button, Tooltip } from "@react-elf/ui";
-import { FaBeer, FaCopy, FaEdit, FaMortarPestle } from "react-icons/fa";
-import { MdContentCopy, MdEdit, MdMore } from "react-icons/md";
+import { ActionGroup, Button } from "@react-elf/ui";
+import { MdContentCopy, MdEdit, MdMore, MdPadding } from "react-icons/md";
 
 const meta = {
   // title: "Components/Actions/Button",
@@ -21,18 +20,18 @@ export const Default: Story = {
     children: (
       <>
         <Button>
-          <FaEdit />
+          <MdEdit />
           Edit
         </Button>
         <Button>
-          <FaCopy />
+          <MdContentCopy />
           Copy
         </Button>
         <Button iconOnly>
-          <FaMortarPestle />
+          <MdPadding />
         </Button>
         <Button quiet>
-          <FaEdit />
+          <MdEdit />
           Edit
         </Button>
       </>
@@ -45,18 +44,18 @@ export const Horizontal: Story = {
     children: (
       <>
         <Button>
-          <FaEdit />
+          <MdEdit />
           Edit
         </Button>
         <Button>
-          <FaCopy />
+          <MdContentCopy />
           Copy
         </Button>
         <Button iconOnly>
-          <FaMortarPestle />
+          <MdPadding />
         </Button>
         <Button quiet>
-          <FaEdit />
+          <MdEdit />
           Edit
         </Button>
       </>
@@ -80,6 +79,350 @@ export const HorizontalQuiet: Story = {
           <MdMore />
         </Button>
         <Button quiet>
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const Vertical: Story = {
+  args: {
+    direction: "vertical",
+    children: (
+      <>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+        <Button>
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly>
+          <MdPadding />
+        </Button>
+        <Button quiet>
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const VerticalQuiet: Story = {
+  args: {
+    direction: "vertical",
+    children: (
+      <>
+        <Button quiet>
+          <MdEdit />
+          Edit
+        </Button>
+        <Button quiet>
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly quiet>
+          <MdMore />
+        </Button>
+        <Button quiet>
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const SizeDefault: Story = {
+  args: {
+    children: (
+      <>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+        <Button>
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly>
+          <MdPadding />
+        </Button>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const SizeSmall: Story = {
+  args: {
+    children: (
+      <>
+        <Button size="small">
+          <MdEdit />
+          Edit
+        </Button>
+        <Button size="small">
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly size="small">
+          <MdPadding />
+        </Button>
+        <Button size="small">
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const SizeLarge: Story = {
+  args: {
+    children: (
+      <>
+        <Button size="large">
+          <MdEdit />
+          Edit
+        </Button>
+        <Button size="large">
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly size="large">
+          <MdPadding />
+        </Button>
+        <Button size="large">
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const SizeExtraLarge: Story = {
+  args: {
+    children: (
+      <>
+        <Button size="extra-large">
+          <MdEdit />
+          Edit
+        </Button>
+        <Button size="extra-large">
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly size="extra-large">
+          <MdPadding />
+        </Button>
+        <Button size="extra-large">
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const Compact: Story = {
+  args: {
+    compact: true,
+    children: (
+      <>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+        <Button>
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly>
+          <MdPadding />
+        </Button>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const Justified: Story = {
+  args: {
+    justified: true,
+    children: (
+      <>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+        <Button>
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly>
+          <MdPadding />
+        </Button>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const JustifiedCompact: Story = {
+  args: {
+    justified: true,
+    compact: true,
+    children: (
+      <>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+        <Button>
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly>
+          <MdPadding />
+        </Button>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const Quiet: Story = {
+  args: {
+    children: (
+      <>
+        <Button quiet>
+          <MdEdit />
+          Edit
+        </Button>
+        <Button quiet>
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly quiet>
+          <MdPadding />
+        </Button>
+        <Button quiet>
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const Selection: Story = {
+  args: {
+    compact: true,
+    children: (
+      <>
+        <Button selected>
+          <MdEdit />
+          Edit
+        </Button>
+        <Button>
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly>
+          <MdPadding />
+        </Button>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const MultiSelection: Story = {
+  args: {
+    children: (
+      <>
+        <Button selected>
+          <MdEdit />
+          Edit
+        </Button>
+        <Button selected>
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly>
+          <MdPadding />
+        </Button>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const Collapsed: Story = {
+  args: {
+    collapsed: true,
+    children: (
+      <>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+        <Button>
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly>
+          <MdPadding />
+        </Button>
+        <Button>
+          <MdEdit />
+          Edit
+        </Button>
+      </>
+    ),
+  },
+};
+
+export const Disabled: Story = {
+  args: {
+    children: (
+      <>
+        <Button disabled>
+          <MdEdit />
+          Edit
+        </Button>
+        <Button disabled>
+          <MdContentCopy />
+          Copy
+        </Button>
+        <Button iconOnly disabled>
+          <MdPadding />
+        </Button>
+        <Button disabled>
           <MdEdit />
           Edit
         </Button>